not good :)

+++ adm/src/main/java/com/sh/adm/user/controller/UserApiController.java
@@ -2,9 +2,9 @@ package com.sh.adm.user.controller;
 
 
 import com.sh.adm.ifs.CrudInterface;
-import com.sh.adm.model.network.Header;
-import com.sh.adm.user.dto.UserApiRequest;
-import com.sh.adm.user.dto.UserApiResponse;
+import com.sh.adm.user.model.network.Header;
+import com.sh.adm.user.model.dto.UserApiRequest;
+import com.sh.adm.user.model.dto.UserApiResponse;
 import com.sh.adm.user.service.UserApiLogicService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -25,7 +25,6 @@ public class UserApiController implements CrudInterface<UserApiRequest, UserApiR
     // Header 공통 -> CrudInterface
     // REST 연동 전 규칙 설정 : @PostMapping, @GetMapping, @PutMapping, @DeleteMapping
     //  -> 공통, 중복 부분 추상화 리팩토링
-
     private final UserApiLogicService userApiLogicService;
 
     @GetMapping("")
@@ -39,29 +38,38 @@ public class UserApiController implements CrudInterface<UserApiRequest, UserApiR
     @ResponseStatus(HttpStatus.CREATED)  // 201
     public Header<UserApiResponse> create(@RequestBody Header<UserApiRequest> request) {
         log.info("{}",request);
-        return userApiLogicService.create(request);
+        UserApiRequest data = request.getData();
+        if(data.getAccount() == null || data.getEmail() == null || data.getPassword() == null)
+            return Header.error("Need a data");
+        return userApiLogicService.create(data);
     }
 
     @Override
     @GetMapping("{id}")   //"/api/user/{ids}"
     public Header<UserApiResponse> read(@PathVariable("id") Long id) {
+        if(id == null) return Header.error("Must be need to parameter");
         return userApiLogicService.read(id);
     }
 
     @Override
     @PutMapping("")
     public Header<UserApiResponse> update(@RequestBody Header<UserApiRequest> request) {
-        return userApiLogicService.update(request);
+        UserApiRequest data = request.getData();
+        if(data.getId()==null || data.getAccount() == null)
+            return Header.error("Need a data");
+        return userApiLogicService.update(data);
     }
 
     @PatchMapping("/change_password/{id}")
     public Header update(@PathVariable Long id, String password) {
+        if(id== null || password == null) return Header.error("Need a data");
         return userApiLogicService.update(id, password);
     }
 
     @Override
     @DeleteMapping("{id}")
     public Header delete(@PathVariable Long id) {
+        if(id == null) return Header.error("Need a data");
         return userApiLogicService.delete(id);
     }
